// src/ai/solver.d.ts
export interface IModelVariableConstraint {
  /** The variable should be grater or equal to this value. */
  min?: number;
  /** The variable should be less or equal to this value. */
  max?: number;
  /** The variable should be equal to this value. */
  equal?: number;
}

/**
* Specifies the options when solving the problem.
*/
export interface IModelOptions {
  /**
   * For MILP problems, specifies the relative tolerance of the objective,
   * where `0` means 0% and `1` means 100%.
   */
  tolerance?: number;
  /**
   * How many milliseconds you want to allow for the solver to try
   * and solve the model you're running.
   */
  timeout?: number;
  /**
   * Use MIR cuts.
   * @deprecated NOT WORKING
   */
  useMIRCuts?: boolean;
  /**
   * Defaults to `true`.
   */
  exitOnCycles?: boolean;
}

/**
* Represents an LP/MILP problem.
* @typeparam TSolutionVar the decision variables that will be outputed to the `Solution` object.
* @typeparam TInternalVar the decision variables that will not be outputed to the `Solution` object.
*/
export interface IModel<TSolutionVar extends string = string, TInternalVar extends string = string> {
  /** Name of the variable that will be the optimization objective. */
  optimize: (TSolutionVar | TInternalVar);
  /** To which direction to optimize the objective. */
  opType: "max" | "min";
  /**
   * Optimization constraints.
   * Specify an object with variable name as keys.
  */
  constraints: { [variable in (TSolutionVar | TInternalVar)]?: IModelVariableConstraint };
  /**
   * Variable identity relations.
   * Specify an object with variable name as keys. These variables will be outputted into solution.
   * The values of the object represents a linear combination of all the (rest of) variables.
   * @example
   *      ```
   *      {
   *          x: { x1: 10, x2: 5, x3: 2, x: 1 }       // x = 10 x1 + 5 x2 + 2 x3
   *      }
   *      ```
   */
  variables: { [variable in TSolutionVar]?: { [variable in (TSolutionVar | TInternalVar)]?: number } };
  /**
   * For each variable in the MILP problem, specifies whether it is an integer variable.
   * You need to specify `true` or `1` for integer variable.
   * If not specified, all the variables are continual non-negative (range `[0,+∞)`).
   */
  ints?: { [variable in (TSolutionVar | TInternalVar)]?: boolean | 0 | 1 };
  /**
   * For each variable in the MILP problem, specifies whether it is a binary variable.
   * You need to specify `true` or `1` for binary variable.
   * If not specified, all the variables are continual non-negative (range `[0,+∞)`).
   */
  binaries?: { [variable in (TSolutionVar | TInternalVar)]?: boolean | 0 | 1 };
  /**
   * For each variable in the MILP problem, specifies whether it is an unrestricted variable (range `(-∞,+∞)`).
   * You need to specify `true` or `1` for unrestricted variable.
   * If not specified, all the variables are continual non-negative (range `[0,+∞)`).
   */
  unrestricted?: { [variable in (TSolutionVar | TInternalVar)]?: boolean | 0 | 1 };
  /**
   * Options for solving this problem.
   */
  options?: IModelOptions;
}